app/requirements.txt
fastapi
uvicorn
scikit-learn
pandas
google-cloud-storage
